+++ Map/DiamondSquare.h
@@ -9,11 +9,11 @@ public:
     ~DiamondSquare();
 
 private:
-    void diamondSquare ( World*world, float scale );
+    void diamondSquare ( World*world, double scale );
     void init ( World *world );
-    void diamondStep ( World* world, int y, int x, float scale );
-    void squareStep ( World *world, int y, int x, float scale );
-    float rangedRandom();
+    void diamondStep ( World* world, int y, int x, double scale );
+    void squareStep ( World *world, int y, int x, double scale );
+    double rangedRandom();
     void setDepth ( World *world );
     bool  checkIndex ( int index,int size );
     int step;
